+++ View/Converters/HourConverter.cs
@@ -5,12 +5,12 @@ using System.Windows.Data;
 
 namespace View.Converters
 {
-    [ValueConversion(typeof(int), typeof(string))]
+    [ValueConversion(typeof(byte), typeof(string))]
     internal sealed class HourConverter : IValueConverter
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            var hour = (int)value;
+            var hour = (byte)value;
             var result = System.Convert.ToString(hour);
             if (result.Length < 2) result = $"0{result}";
 
@@ -19,7 +19,7 @@ namespace View.Converters
 
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            var valid = int.TryParse((string)value, out var hour);
+            var valid = byte.TryParse((string)value, out var hour);
             if (!valid) return DependencyProperty.UnsetValue;
             if (hour is > 23 or < 0) hour = 23;
